Integrate review comments
