+++ token-swap/program/src/state.rs
@@ -7,7 +7,11 @@ use solana_sdk::pubkey::Pubkey;
 #[repr(C)]
 #[derive(Clone, Copy, Debug, Default, PartialEq)]
 pub struct SwapInfo {
-    /// Nonce used in construction of program_address
+    /// Nonce used in program address.
+    /// The program address is created deterministically with the nonce,
+    /// swap program id, and swap account pubkey.  This program address has
+    /// authority over the swap's token A account, token B account, and pool
+    /// token mint.
     pub nonce: u8,
     /// Token A
     /// The Liquidity token is issued against this value.
@@ -59,5 +63,46 @@ impl Invariant {
     /// Exchange rate
     pub fn exchange_rate(&self, token_a: u64) -> Option<u64> {
         token_a.checked_mul(self.token_b)?.checked_div(self.token_a)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::mem::size_of;
+
+    #[test]
+    fn test_state_swap_info_deserialization() {
+        let nonce = 255;
+        let token_a_raw = [1u8; 32];
+        let token_b_raw = [2u8; 32];
+        let pool_mint_raw = [3u8; 32];
+        let token_a = Pubkey::new_from_array(token_a_raw);
+        let token_b = Pubkey::new_from_array(token_b_raw);
+        let pool_mint = Pubkey::new_from_array(pool_mint_raw);
+        let numerator = 1;
+        let denominator = 4;
+        let fee = Fee { numerator, denominator };
+        let state = State::Init(SwapInfo { nonce, token_a, token_b, pool_mint, fee, });
+
+        let mut data = [0u8; size_of::<State>()];
+        state.serialize(&mut data).unwrap();
+        let deserialized = State::deserialize(&data).unwrap();
+        assert_eq!(state, deserialized);
+
+        let mut data = vec![];
+        data.push(1 as u8);
+        data.push(nonce);
+        data.extend_from_slice(&token_a_raw);
+        data.extend_from_slice(&token_b_raw);
+        data.extend_from_slice(&pool_mint_raw);
+        data.extend_from_slice(&[0u8; 7]); // padding
+        data.push(denominator as u8);
+        data.extend_from_slice(&[0u8; 7]); // padding
+        data.push(numerator as u8);
+        data.extend_from_slice(&[0u8; 7]); // padding
+        data.extend_from_slice(&[0u8; 7]); // padding
+        let deserialized = State::deserialize(&data).unwrap();
+        assert_eq!(state, deserialized);
     }
 }
